Moved tests to the test_authentication.py
